Split domain families track processors
