+++ app/assets/javascripts/3dbio_viewer/src/data/repositories/protvista/tracks/structure-coverage.ts
@@ -1,6 +1,6 @@
 import _ from "lodash";
 import { Fragments, getFragments } from "../../../../domain/entities/Fragment2";
-import { Track } from "../../../../domain/entities/Track";
+import i18n from "../../../../domain/utils/i18n";
 import { config } from "../config";
 import { subtracks } from "../definitions";
 
@@ -9,17 +9,11 @@ export interface Coverage {
 }
 
 export function getStructureCoverageFragments(coverage: Coverage): Fragments {
-    const itemKey = "region";
-    const trackConfig = config.tracks[itemKey];
-    const name = "Region";
-    const items = coverage["Structure coverage"];
-    if (!items) return [];
-
     return getFragments(coverage["Structure coverage"] || [], item => ({
         subtrack: subtracks.structureCoverage,
         start: item.start,
         end: item.end,
-        description: "Sequence segment covered by the structure",
-        color: config.colorByTrackName[itemKey],
+        description: i18n.t("Sequence segment covered by the structure"),
+        color: config.colorByTrackName.region,
     }));
 }
