+++ app/assets/javascripts/3dbio_viewer/src/data/repositories/protvista/utils.ts
@@ -1,3 +1,5 @@
+import _ from "lodash";
+
 export function getName(s: string | undefined) {
     return (s || "").replace(/_/g, " ");
 }
@@ -7,4 +9,23 @@ export function getId(name: string): string {
         .replace("&", "and")
         .replace(/[^\w]+/g, "-")
         .toLowerCase();
+}
+
+export interface Item {
+    name: string | undefined;
+    values: Array<string | undefined> | undefined;
+}
+
+export function getStringFromItems(items: Item[]): string {
+    return _(items)
+        .map(item =>
+            item.values && !_.isEmpty(item.values)
+                ? _.compact([
+                      item.name ? `<b style="color: grey">${item.name}:</b>` : undefined,
+                      item.name ? item.values.join(", ") : `<b>${item.values.join(", ")}</b>`,
+                  ]).join(" ")
+                : null
+        )
+        .compact()
+        .join("<br/>");
 }
